+++ src/utility/river-service.ts
@@ -1,5 +1,4 @@
 // src/utility/data-service.ts
-// Define interfaces for the data structures based on data.txt
 export interface RiverLevel {
   id: string;
   value: number;
